+++ ruma-events-macros/src/gen.rs
@@ -153,7 +153,7 @@ impl ToTokens for RumaEvent {
                         }
 
                         /// Additional key-value pairs not signed by the homeserver.
-                        fn unsigned(&self) -> &std::collections::BTreeMap<String, serde_json::Value> {
+                        fn unsigned(&self) -> &ruma_events::UnsignedData {
                             &self.unsigned
                         }
                     }
@@ -295,8 +295,8 @@ fn populate_room_event_fields(content_name: Ident, fields: Vec<Field>) -> Vec<Fi
         pub sender: ruma_identifiers::UserId,
 
         /// Additional key-value pairs not signed by the homeserver.
-        #[serde(default, skip_serializing_if = "std::collections::BTreeMap::is_empty")]
-        pub unsigned: std::collections::BTreeMap<String, serde_json::Value>,
+        #[serde(skip_serializing_if = "ruma_events::UnsignedData::is_empty")]
+        pub unsigned: ruma_events::UnsignedData,
     };
 
     fields.extend(punctuated_fields.into_iter().map(|p| p.field));