+++ src/room/member.rs
@@ -230,7 +230,7 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use crate::util::serde_json_eq_try_from_raw;
+    use crate::{util::serde_json_eq_try_from_raw, UnsignedData};
 
     #[test]
     fn serde_with_no_prev_content() {
@@ -247,7 +247,7 @@ mod tests {
             room_id: Some(RoomId::try_from("!n8f893n9:example.com").unwrap()),
             sender: UserId::try_from("@carl:example.com").unwrap(),
             state_key: "example.com".to_string(),
-            unsigned: BTreeMap::new(),
+            unsigned: UnsignedData::default(),
             prev_content: None,
         };
         let json = json!({
@@ -279,7 +279,7 @@ mod tests {
             room_id: Some(RoomId::try_from("!n8f893n9:example.com").unwrap()),
             sender: UserId::try_from("@carl:example.com").unwrap(),
             state_key: "example.com".to_string(),
-            unsigned: BTreeMap::new(),
+            unsigned: UnsignedData::default(),
             prev_content: Some(MemberEventContent {
                 avatar_url: None,
                 displayname: None,
@@ -335,7 +335,7 @@ mod tests {
             room_id: Some(RoomId::try_from("!jEsUZKDJdhlrceRyVU:example.org").unwrap()),
             sender: UserId::try_from("@alice:example.org").unwrap(),
             state_key: "@alice:example.org".to_string(),
-            unsigned: BTreeMap::new(),
+            unsigned: UnsignedData::default(),
             prev_content: None,
         };
         let json = json!({
@@ -390,7 +390,7 @@ mod tests {
             room_id: Some(RoomId::try_from("!jEsUZKDJdhlrceRyVU:example.org").unwrap()),
             sender: UserId::try_from("@alice:example.org").unwrap(),
             state_key: "@alice:example.org".to_string(),
-            unsigned: BTreeMap::new(),
+            unsigned: UnsignedData::default(),
             prev_content: Some(MemberEventContent {
                 avatar_url: Some("mxc://example.org/SEsfnsuifSDFSSEF".to_owned()),
                 displayname: Some("Alice Margatroid".to_owned()),